feat: db관련 의존성 추가
