+++ docs/2주차 기능명세서.md
@@ -1,0 +1,85 @@
+# 2주차 기능 명세서
+
+## [요구사항]
+
+---
+
+### 1.입력
+> CLI (Command linke inteface)한 환경에서 실행되도록 한다.
+
+[ 프로그램 실행 구문 ]
+```
+=== Manage Program ===
+Type customer to select customer menu.
+Type voucher to select customer menu.
+Type exit to exit the program.
+```
+
+[ voucher 실행 구문 ]
+```
+=== Voucher Manage ===
+Type exit to exit the program.
+Type create to create a new voucher.
+Type list to list all vouchers.
+```
+
+[ customer 실행 구문 ]
+```
+=== Customer Manage ===
+Type exit to exit the program.
+Type create to create a new customer.
+Type list to list all customers.
+```
+
+### 입력 예외 추가 기능
+- 입력시 `양쪽 공백`은 `삭제`
+- `대소문자 상관없이` 동일한 메뉴 이름이면 사용가능
+
+### 2. 회원
+> `create`와 `list` 커맨드를 통해 회원를 생성 및 조회할 수 있다.
+
+- `create`: 회원 생성
+- `list`: 회원 조회
+- customers 테이블 정의 및 추가
+    - CustomerRepository 추가 및 `JdbcTemplate`을 사용해서 구현
+    - CRUD 모든 기능 추가
+
+
+## 3. 바우처
+- `JdbcTemplate`을 사용해서 구현
+-  CRUD 모든 기능 추가
+
+## 4. 바우처 지갑 (심화)
+- 특정 고객에게 바우처를 할당할 수 있습니다.
+- 고객이 어떤 바우처를 보유하고 있는지 조회할 수 있어야 합니다.
+- 고객이 보유한 바우처를 제거할 수 있어야 합니다.
+- 특정 바우처를 보유한 고객을 조회할 수 있어야 합니다.
+
+<br>
+
+## 기능
+
+---
+
+### 1. 입력
+- [ ] 회원과 바우처 선택하는 기능
+
+### 2. 바우처
+- [ ] JdbcTemplate를 사용하여 Repostiroy 관리
+- [ ] CRUD 기능
+
+### 3. 회원
+- [ ] 회원 CRUD 기능
+  - [ ] 회원 추가
+  - [ ] 회원 삭제
+  - [ ] 회원 업데이트
+  - [ ] 회원 조회
+- [ ] JdbcTemplate를 사용하여 Repository 관리
+
+### 4. 심화
+- 특정 고객에게 바우처를 할당할 수 있습니다.
+- 고객이 어떤 바우처를 보유하고 있는지 조회할 수 있어야 합니다.
+- 고객이 보유한 바우처를 제거할 수 있어야 합니다.
+- 특정 바우처를 보유한 고객을 조회할 수 있어야 합니다.
+
+
