+++ docs/README.md
@@ -34,7 +34,8 @@ Type list to list all vouchers.
   - 파일을 읽을 수 있고 리스트 조회 가능
 
 
-### 3. 기타 요구사항
+
+### 기타 요구사항
 - `YAML`프러퍼티를 만들고 어떤 설정을 만들 수 있을지 
 - 실행가능한 `jar`파일을 생성
 - `logback`을 통한 에러 파일로 기록
