+++ src/test/java/org/weekly/weekly/voucher/VoucherRepositoryTest.java
@@ -1,0 +1,9 @@
+package org.weekly.weekly.voucher;
+
+
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("test")
+public class VoucherRepositoryTest {
+
+}
